+++ packages/datafile-manager/src/httpPollingDatafileManager.ts
@@ -20,7 +20,6 @@ import { DatafileManager, DatafileManagerConfig, DatafileUpdate } from './datafi
 import EventEmitter from './eventEmitter'
 import { AbortableRequest, Response, Headers } from './http';
 import { DEFAULT_UPDATE_INTERVAL, MIN_UPDATE_INTERVAL, DEFAULT_URL_TEMPLATE } from './config'
-import { TimeoutFactory, DEFAULT_TIMEOUT_FACTORY } from './timeoutFactory'
 import BackoffController from './backoffController';
 
 const logger = getLogger('DatafileManager')
@@ -61,7 +60,7 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
 
   private readonly updateInterval: number
 
-  private cancelTimeout?: () => void
+  private currentTimeout: any
 
   private isStarted: boolean
 
@@ -69,11 +68,15 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
 
   private datafileUrl: string
 
-  private timeoutFactory: TimeoutFactory
-
-  private currentRequest?: AbortableRequest
+  private currentRequest: AbortableRequest | null
 
   private backoffController: BackoffController
+
+  // When true, this means the update interval timeout fired before the current
+  // sync completed. In that case, we should sync again immediately upon
+  // completion of the current request, instead of waiting another update
+  // interval.
+  private syncOnCurrentRequestComplete: boolean
 
   constructor(config: DatafileManagerConfig) {
     const configWithDefaultsApplied: DatafileManagerConfig = {
@@ -84,7 +87,6 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
       datafile,
       autoUpdate = false,
       sdkKey,
-      timeoutFactory = DEFAULT_TIMEOUT_FACTORY,
       updateInterval = DEFAULT_UPDATE_INTERVAL,
       urlTemplate = DEFAULT_URL_TEMPLATE,
     } = configWithDefaultsApplied
@@ -108,7 +110,6 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
 
     this.datafileUrl = sprintf(urlTemplate, sdkKey)
 
-    this.timeoutFactory = timeoutFactory
     this.emitter = new EventEmitter()
     this.autoUpdate = autoUpdate
     if (isValidUpdateInterval(updateInterval)) {
@@ -117,7 +118,10 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
       logger.warn('Invalid updateInterval %s, defaulting to %s', updateInterval, DEFAULT_UPDATE_INTERVAL)
       this.updateInterval = DEFAULT_UPDATE_INTERVAL
     }
+    this.currentTimeout = null
+    this.currentRequest = null
     this.backoffController = new BackoffController()
+    this.syncOnCurrentRequestComplete = false
   }
 
   get(): object | null {
@@ -136,16 +140,16 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
   stop(): Promise<void> {
     logger.debug('Datafile manager stopped')
     this.isStarted = false
-    if (this.cancelTimeout) {
-      this.cancelTimeout()
-      this.cancelTimeout = undefined
+    if (this.currentTimeout) {
+      clearTimeout(this.currentTimeout)
+      this.currentTimeout = null
     }
 
     this.emitter.removeAllListeners()
 
     if (this.currentRequest) {
       this.currentRequest.abort()
-      this.currentRequest = undefined
+      this.currentRequest = null
     }
 
     return Promise.resolve()
@@ -209,15 +213,17 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
       return
     }
 
-    this.currentRequest = undefined
+    this.currentRequest = null
 
-    if (this.autoUpdate) {
-      this.scheduleNextUpdate()
-    }
     if (!this.isReadyPromiseSettled && !this.autoUpdate) {
       // We will never resolve ready, so reject it
       this.rejectReadyPromise(new Error('Failed to become ready'))
     }
+
+    if (this.autoUpdate && this.syncOnCurrentRequestComplete) {
+      this.syncDatafile()
+    }
+    this.syncOnCurrentRequestComplete = false
   }
 
   private syncDatafile(): void {
@@ -241,6 +247,10 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
     this.currentRequest.responsePromise
       .then(onRequestResolved, onRequestRejected)
       .then(onRequestComplete, onRequestComplete)
+
+    if (this.autoUpdate) {
+      this.scheduleNextUpdate()
+    }
   }
 
   private resolveReadyPromise(): void {
@@ -257,8 +267,12 @@ export default abstract class HTTPPollingDatafileManager implements DatafileMana
     const currentBackoffDelay = this.backoffController.getDelay()
     const nextUpdateDelay = Math.max(currentBackoffDelay, this.updateInterval)
     logger.debug('Scheduling sync in %s ms', nextUpdateDelay)
-    this.cancelTimeout = this.timeoutFactory.setTimeout(() => {
-      this.syncDatafile()
+    this.currentTimeout = setTimeout(() => {
+      if (this.currentRequest) {
+        this.syncOnCurrentRequestComplete = true
+      } else {
+        this.syncDatafile()
+      }
     }, nextUpdateDelay)
   }